Extracted ErrorState from TracingStatus; Removed TracingStatus and related classes.
