--- sdk/src/main/java/org/coralibre/android/sdk/InfectionStatus.java
@@ -1,17 +1,0 @@
-/*
- * Copyright (c) 2020 Ubique Innovation AG <https://www.ubique.ch>
- *
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- *
- * SPDX-License-Identifier: MPL-2.0
- */
-
-package org.coralibre.android.sdk;
-
-public enum InfectionStatus {
-	HEALTHY,
-	EXPOSED,
-	INFECTED;
-}
