+++ sdk/src/main/java/org/coralibre/android/sdk/internal/error/old/ErrorHelper.java
@@ -7,7 +7,7 @@
  *
  * SPDX-License-Identifier: MPL-2.0
  */
-package org.coralibre.android.sdk.internal;
+package org.coralibre.android.sdk.internal.error.old;
 
 import android.Manifest;
 import android.bluetooth.BluetoothAdapter;
@@ -22,9 +22,9 @@ import java.util.Collection;
 import java.util.HashSet;
 import java.util.Set;
 
-import org.coralibre.android.sdk.TracingStatus.ErrorState;
 import org.coralibre.android.sdk.internal.bluetooth.BluetoothServiceStatus;
 import org.coralibre.android.sdk.internal.util.LocationServiceUtil;
+
 
 public class ErrorHelper {
